Proyecto5-RAG/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
azure-search-documents==11.4.0
azure-identity==1.16.0
openai==1.30.1
pydantic==2.7.1
langchain-text-splitters==0.2.1
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
